requirements.txt
streamlit
yfinance
openai
langchain
crewai
python-dotenv
